Añadí nuevas carpetas para los dao y dto, y la mayor parte de los cambios es la parte de la vista/interfaz

+++ src/main/java/com/a22ivancp/accesoadatos/model/entities/Subject.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class Subject {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idSubject;
+    private Integer idSubject;
     private String name;
 
     @ManyToMany(mappedBy = "subjects")
